requirements.txt
absl-py==1.0.0
alabaster==0.7.13
anyio==3.7.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
bleach==6.0.0
brax==0.0.15
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
chex==0.1.5
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
contourpy==1.0.7
cookiecutter==2.1.1
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dm-env==1.6
dm-haiku==0.0.5
dm-tree==0.1.8
docker-pycreds==0.4.0
docutils==0.18.1
etils==1.3.0
exceptiongroup==1.1.1
executing==1.2.0
Farama-Notifications==0.0.4
fastjsonschema==2.17.1
filelock==3.12.2
flake8==6.0.0
flax==0.6.0
fonttools==4.39.4
fqdn==1.5.1
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.9
grpcio==1.54.2
gym==0.23.1
gym-notices==0.0.8
gymnasium==0.28.1
gymnasium-robotics==1.2.2
identify==2.5.24
idna==3.4
imageio==2.31.0
imagesize==1.4.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipympl==0.9.3
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
jax==0.3.17
jax-jumpy==1.0.0
jaxlib==0.3.15
jedi==0.18.2
Jinja2==3.0.3
jinja2-time==0.2.0
jmp==0.0.4
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jumanji==0.3.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
msgpack==1.0.5
mujoco==2.3.3
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
nodeenv==1.8.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.22.3
opt-einsum==3.3.0
optax==0.1.4
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pettingzoo==1.23.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
pre-commit==3.3.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.19.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pygame==2.4.0
Pygments==2.15.1
PyOpenGL==3.1.7
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytinyrenderer==0.0.14
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
-e git+https://github.com/gardarjuto/MA-QDax.git@25fcb85bc4c2ebfc58c85886033b5edf5bb15003#egg=qdax
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==11.2.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.2
sentry-sdk==1.25.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==6.2.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
tabulate==0.9.0
tensorboardX==2.6
tensorflow-probability==0.15.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
trimesh==3.22.0
typing_extensions==4.3.0
tzdata==2023.3
uri-template==1.2.0
urllib3==2.0.3
virtualenv==20.23.0
wandb==0.15.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
widgetsnbextension==4.0.7
zipp==3.15.0
